Add files via upload
This project is to develop a system by creating three classes for the vehicle to display detail, book, purchase, and sell the vehicle. My IDE for this coursework was BlueJ. My coursework consists of three classes and they are Vehicle, AutoRickshaw, and ElectricScooter. Here Vehicle class was my parent class and AutoRickshaw and ElectricScooter were sub-classes of Vehicle class. Different methods were created during this coursework and accessor methods were assigned for each class and the various display methods were created to display the data of every class.